Eye vision rework.
Changed the vision, power, and self damage on the eye to use a Tier system with stacking  conditions.

Needs to be fixed as the lower tiers are not triggering but otherwise works.

Safety net of disabling when low on HP, will not kill itself but vision goes down and cost of power goes right up