Air and siege + Chaos
Changed prices on medi vehicles to be a little less annoying.

Increased HP on some aircraft as they were just made of paper.

Reworked Chaos elemental to be 2 less range and have some more range of weapons. should now come in vision of towers and hopefully need to take hits to attack not just float outside of your base

made the Chaos orbs weapons and armaments easier to read